+++ testapp/settings.py
@@ -80,3 +80,35 @@ OIDC_CALLBACK_CLASS = "mozilla_django_oidc_db.views.OIDCCallbackView"
 LOGIN_REDIRECT_URL = reverse_lazy("admin:index")
 
 STATIC_URL = "/static/"
+
+
+# Setup Configuration Settings
+
+IDENTITY_PROVIDER = "https://keycloak.local/realms/digid/"
+
+ADMIN_OIDC_OIDC_RP_CLIENT_ID = "client-id"
+ADMIN_OIDC_OIDC_RP_CLIENT_SECRET = "secret"
+ADMIN_OIDC_OIDC_RP_SCOPES_LIST = ["open_id", "email", "profile", "extra_scope"]
+ADMIN_OIDC_OIDC_RP_SIGN_ALGO = "RS256"
+ADMIN_OIDC_OIDC_RP_IDP_SIGN_KEY = "key"
+ADMIN_OIDC_OIDC_OP_DISCOVERY_ENDPOINT = None
+ADMIN_OIDC_OIDC_OP_JWKS_ENDPOINT = f"{IDENTITY_PROVIDER}protocol/openid-connect/certs"
+ADMIN_OIDC_OIDC_OP_AUTHORIZATION_ENDPOINT = (
+    f"{IDENTITY_PROVIDER}protocol/openid-connect/auth"
+)
+ADMIN_OIDC_OIDC_OP_TOKEN_ENDPOINT = f"{IDENTITY_PROVIDER}protocol/openid-connect/token"
+ADMIN_OIDC_OIDC_OP_USER_ENDPOINT = (
+    f"{IDENTITY_PROVIDER}protocol/openid-connect/userinfo"
+)
+ADMIN_OIDC_USERNAME_CLAIM = ["claim_name"]
+ADMIN_OIDC_GROUPS_CLAIM = ["groups_claim_name"]
+ADMIN_OIDC_CLAIM_MAPPING = {"first_name": "given_name"}
+ADMIN_OIDC_SYNC_GROUPS = False
+ADMIN_OIDC_SYNC_GROUPS_GLOB_PATTERN = "local.groups.*"
+ADMIN_OIDC_DEFAULT_GROUPS = ["Admins", "Read-only"]
+ADMIN_OIDC_MAKE_USERS_STAFF = True
+ADMIN_OIDC_SUPERUSER_GROUP_NAMES = ["superuser"]
+ADMIN_OIDC_OIDC_USE_NONCE = False
+ADMIN_OIDC_OIDC_NONCE_SIZE = 48
+ADMIN_OIDC_OIDC_STATE_SIZE = 48
+ADMIN_OIDC_USERINFO_CLAIMS_SOURCE = "id_token"