+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 0.20.0 (????)
 =============
 
+New Features:
+
+* Add optional support for ``django setup configuration``
 
 
 0.19.0 (2024-07-02)