+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+0.20.0 (????)
+=============
+
+New Features:
+
+* Add optional support for ``django setup configuration``
+
+
 0.19.0 (2024-07-02)
 ===================